+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # VocalFusion Raspberry Pi Setup Change Log
 
+## 1.3.0
+
+  * Updated to AVS Device SDK v1.6. [See details here](https://github.com/xmos/avs-device-sdk/blob/xmos_v1.6/CHANGELOG.md).
+  * Using vocalfusion-rpi-setup v1.2. [See details here](https://github.com/xmos/vocalfusion-rpi-setup/blob/v1.2/CHANGELOG.md).
+
 ## 1.2.0
 
   * Using vocalfusion-rpi-setup v1.1. [See details here](https://github.com/xmos/vocalfusion-rpi-setup/blob/v1.1/CHANGELOG.md).